UI buttons are working!!!!! & the player cant double jump anymore but :
i couldnt manage to use the feet, because when we use feet we look at the distance bettwin the feet and the ground, and the ground in oure case is huge and difigned square, so it dousent look at the actual tile map that has a collider.
so for now i just wrote that if the player's velocity is 0 he can jump.
not the best but works. most of the time.